created separate folders to organize my components based on similar functionality/purpose for better structure and maintainability
