+++ client/src/components/Root.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Navbar from './Navbar';
+import Navbar from './layout/Navbar';
 
 export default function Root() {
   return (
