+++ client/src/components/layout/Sidebar.jsx
@@ -1,9 +1,9 @@
 import { useContext } from 'react';
 import { NavLink, Link } from 'react-router-dom';
 import { Context } from '../../context/Context';
-import ToggleThemeButton from '../ToggleThemeButton';
+import ToggleThemeButton from '../ui/ToggleThemeButton';
 
-import logo from '../../public/assets/logo.svg';
+import logo from '../../../public/assets/logo.svg';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faArrowRightFromBracket,
