+++ client/src/components/profile/DefaultAvatar.jsx
