+++ client/src/components/ui/Modal.jsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { Context } from '../context/Context';
+import { Context } from '../../context/Context';
 
 export default function Modal({ children }) {
   const { closeModal } = useContext(Context);
